Check for email_session cookie on API endpoints.

+++ routes/api.js
@@ -9,22 +9,25 @@ module.exports = ( router ) => {
   }));
 
   router.get( '/emails/:userId', ( req, res ) => {
-    const { userId } = req.params;
+    if( req.cookies.email_session ) {
+      const { userId } = req.params;
 
-    db.any( `SELECT * from emails WHERE "recipientId"=${ userId } ORDER BY id DESC` )
-      .then( results => {
-        const posts = results.slice(0);
-        return posts
-          ? res.status( 200 ).json( posts )
-          : res.status( 403 ).send
-      })
-      .catch( err => {
-        return res.status( 500 ).send();
-      })
+      db.any( `SELECT * from emails WHERE "recipientId"=${ userId } ORDER BY id DESC` )
+        .then( results => {
+          const posts = results.slice(0);
+          return posts
+            ? res.status( 200 ).json( posts )
+            : res.status( 403 ).send
+        })
+        .catch( err => {
+          return res.status( 500 ).send();
+        })
+    }
   });
 
   router.post( '/sendEmail', ( req, res ) => {
-    const { userId, recipient, subject, message } = req.body;
+    if( req.cookies.email_session ) {
+      const { userId, recipient, subject, message } = req.body;
 
     let messageData = {
       recipient: recipient,
@@ -66,33 +69,38 @@ module.exports = ( router ) => {
       }).catch( error => {
         console.log('FETCH AUTHOR ID ERROR', error);
       });
-  })
+    }
+  });
 
   router.post( '/updateImportance', ( req, res ) => {
-    const { messageId } = req.body;
+    if( req.cookies.email_session ) {
+      const { messageId } = req.body;
 
-    db.any( `UPDATE emails
-      SET isimportant = NOT isimportant
-      WHERE id = ${ messageId }` )
-      .then( () => res.status( 200 ).send( 'Message was updated successfully' ) )
-      .catch( error => {
-        console.log( 'MARK AS IMPORTANT ERROR: ', error );
+      db.any( `UPDATE emails
+        SET isimportant = NOT isimportant
+        WHERE id = ${ messageId }` )
+        .then( () => res.status( 200 ).send( 'Message was updated successfully' ) )
+        .catch( error => {
+          console.log( 'MARK AS IMPORTANT ERROR: ', error );
 
-        res.status( 500 ).send( 'Something went wrong...Please try again' );
-      })
-  })
+          res.status( 500 ).send( 'Something went wrong...Please try again' );
+        })
+    }
+  });
 
   router.post( '/deleteEmail', ( req, res ) => {
-    const { messageId } = req.body;
+    if( req.cookies.email_session ) {
+      const { messageId } = req.body;
 
-    db.any( `DELETE FROM emails WHERE id=${ messageId }` )
-      .then( () => res.status( 200 ).send( 'The message was deleted successfully' ) )
-      .catch( error => {
-        console.log( 'DELTE EMAIL ERROR: ', error );
+      db.any( `DELETE FROM emails WHERE id=${ messageId }` )
+        .then( () => res.status( 200 ).send( 'The message was deleted successfully' ) )
+        .catch( error => {
+          console.log( 'DELTE EMAIL ERROR: ', error );
 
-        res.status( 500 ).send( 'Something went wrong...Please try again' );
-      });
-  })
+          res.status( 500 ).send( 'Something went wrong...Please try again' );
+        });
+    }
+  });
 
   return router;
 }
